test(#285): Add Tests for z.number().int()

+++ packages/kotti-ui/source/kotti-navbar/types.ts
@@ -3,7 +3,7 @@ import { z } from 'zod'
 
 export namespace KottiNavbar {
 	export const notificationSchema = z.object({
-		count: z.number(),
+		count: z.number().int(),
 		link: z.string(),
 		title: z.string(),
 	})
